requirements.txt
python-pptx==0.6.18
python-docx==0.8.11